+++ flecsi/exec/test/future.cc
@@ -18,17 +18,17 @@ init(double a, double_field::accessor<wo> ga) {
 }
 
 int
-check(future<double> x, double_field::accessor<ro> ga) {
+check(exec::cpu s, future<double> x, double_field::accessor<ro> ga) {
   UNIT("TASK") {
     static_assert(std::is_same_v<decltype(ga[0]), const double &>);
-    EXPECT_EQ(x.get(), ga[0] + 1 + color());
+    EXPECT_EQ(x.get(), ga[0] + 1 + s.launch().index);
     EXPECT_EQ(ga[0], -ga[1]);
   };
 }
 
 double
-index_init(double a, exec::launch_domain) {
-  return a + color();
+index_init(exec::cpu s, double a, exec::launch_domain) {
+  return a + s.launch().index;
 }
 
 void
@@ -43,8 +43,8 @@ index_void_task(exec::launch_domain) {
 } // namespace
 
 int
-reduction_task(int a, exec::launch_domain) {
-  return a + color();
+reduction_task(exec::cpu s, int a, exec::launch_domain) {
+  return a + s.launch().index;
 }
 
 int
@@ -58,17 +58,17 @@ future_driver() {
     // single future
     auto f = execute<init>(d, energy);
 
-    EXPECT_EQ(test<check>(f, energy), 0);
+    EXPECT_EQ(test<check>(exec::on, f, energy), 0);
     EXPECT_EQ(f.get(), ++d);
 
     // future map
     const exec::launch_domain ld{run::context::instance().processes()};
-    auto fm = execute<index_init>(d, ld);
+    auto fm = execute<index_init>(exec::on, d, ld);
     for(auto v : fm.all())
       EXPECT_EQ(v, d++);
 
     // For all values because it's an index future:
-    EXPECT_EQ(test<check>(fm, energy), 0);
+    EXPECT_EQ(test<check>(exec::on, fm, energy), 0);
 
     auto fv = execute<void_task>();
 
@@ -81,13 +81,13 @@ future_driver() {
 
     int a = 7;
     // checking reduction operations
-    auto fmin = reduce<reduction_task, exec::fold::min>(a, ld);
+    auto fmin = reduce<reduction_task, exec::fold::min>(exec::on, a, ld);
     EXPECT_EQ(fmin.get(), a);
 
-    auto fmax = reduce<reduction_task, exec::fold::max>(a, ld);
+    auto fmax = reduce<reduction_task, exec::fold::max>(exec::on, a, ld);
     EXPECT_EQ(fmax.get(), int(a + run::context::instance().processes() - 1));
 
-    auto fsum = reduce<reduction_task, exec::fold::sum>(a, ld);
+    auto fsum = reduce<reduction_task, exec::fold::sum>(exec::on, a, ld);
     int sum = 0;
     for(Color i = 0; i < run::context::instance().processes(); i++)
       sum += a + i;