+++ tutorial/3-execution/2-index-task.cc
@@ -5,12 +5,12 @@
 
 using namespace flecsi;
 
-// Task with no arguments.
+// Task with special arguments.
 
 void
-task(exec::launch_domain) {
-  flog(info) << "Hello World from color " << color() << " of " << colors()
-             << std::endl;
+task(exec::cpu s, exec::launch_domain) {
+  flog(info) << "Hello World from point task " << s.launch().index << " of "
+             << s.launch().size << std::endl;
 }
 
 // Advance control point.
@@ -19,6 +19,6 @@ void
 advance(control_policy &) {
   exec::launch_domain ld{4};
 
-  execute<task>(ld);
+  execute<task>(exec::on, ld);
 } // advance()
 control::action<advance, cp::advance> advance_action;