+++ doc/sphinx/src/news.rst
@@ -68,6 +68,8 @@ New features
 
   * Task parameters can be of a broader set of types; in particular, they need not be serializable.
   * A task parameter can be a ``std::vector`` of a FleCSI type such as a field accessor or a ``std::tuple`` that includes such a type (recursively).
+  * ``exec::cpu``, ``exec::gpu``, and ``exec::omp`` are task parameter types that indicate an execution ``space`` and provide task-launch information.
+    ``exec::on`` is a token argument to pass for them.
   * Index futures provide ``all`` to get all results.
 
 * Topologies